merge developer from main

+++ src/contexts/AuthContext.tsx
@@ -13,16 +13,20 @@ import { IUserData, IUserLogin, loginUsers } from "../services/loginUserApi";
 import { toast } from "react-toastify";
 import { useNavigate } from "react-router-dom";
 import { IUpdateUser, upDateUserApi } from "../services/updateUserApi";
-/* import { string } from "yup"; */
+import { getUsers } from "../services/getUser";
+import { IUserRegister } from "../services/registerUserApi";
+import { getAnimalsId } from "../services/getAnimalsId";
+import { getSuppliesApi, ISupplies } from "../services/getSuppliesApi";
 
-export interface IAuthContexProps {
+export interface IAuthContextProps {
   children: ReactNode;
 }
 
-interface IAuthContex {
+interface IAuthContext {
   loginUser: (data: IUserLogin) => Promise<void>;
   loginRoute: () => void;
   listAnimals: IAnimals[];
+  listSupplies: ISupplies[];
   user: IUserData;
   isLogged: boolean;
   loading: boolean;
@@ -37,13 +41,21 @@ interface IAuthContex {
   modalUpdateUser: boolean;
   setModalUpdateUser: Dispatch<SetStateAction<boolean>>;
   updateUser: (id: IUpdateUser) => Promise<void>;
+  registerUser: (data: IUserRegister) => void;
+  lisAnimalsUser: IAnimals[];
+  registerPet: (data: {}) => void;
+  setIsShowModalPet: Dispatch<SetStateAction<boolean>>;
+  isOpenModalSupplies: boolean;
+  setIsOpenModalSupplis: Dispatch<SetStateAction<boolean>>;
+  isShowModalPet: boolean;
 }
 
-export const AuthContext = createContext<IAuthContex>({} as IAuthContex);
+export const AuthContext = createContext<IAuthContext>({} as IAuthContext);
 
-const AuthProvider = ({ children }: IAuthContexProps) => {
+const AuthProvider = ({ children }: IAuthContextProps) => {
   const navigate = useNavigate();
-  const [listAnimals, setListAnimals] = useState([]);
+  const [listAnimals, setListAnimals] = useState<IAnimals[]>([]);
+  const [lisAnimalsUser, setLisAnimalsUser] = useState<IAnimals[]>([]);
   const [user, setUser] = useState<IUserData>({} as IUserData);
   const [loading, setLoading] = useState<boolean>(true);
   const [isLogged, setIsLogged] = useState<boolean>(false);
@@ -51,10 +63,13 @@ const AuthProvider = ({ children }: IAuthContexProps) => {
   const [donationButton, setDonationButton] = useState<boolean>(true);
   const [adopted, setAdopted] = useState<boolean>(true);
   const [modalUpdateUser, setModalUpdateUser] = useState<boolean>(false);
+  const [isShowModalPet, setIsShowModalPet] = useState<boolean>(false);
+  const [listSupplies, setListSupplies] = useState<ISupplies[]>([]);
+  const [isOpenModalSupplies, setIsOpenModalSupplis] = useState(false);
 
   useEffect(() => {
     getAnimals();
-  }, [adopted]);
+  }, []);
 
   useEffect(() => {
     const loadUser = async () => {
@@ -64,6 +79,17 @@ const AuthProvider = ({ children }: IAuthContexProps) => {
           api.defaults.headers = {
             Authorization: `bearer ${token}`,
           } as ICommonHeaderProperties;
+
+          getUsers().then((res: any) => {
+            setUser(res);
+          });
+
+          getAnimalsId().then((res) => setLisAnimalsUser(res));
+
+          getSuppliesApi().then((res) => {
+            setListSupplies(res);
+          });
+
           setIsLogged(true);
         } catch (err) {
           console.log(err);
@@ -72,33 +98,37 @@ const AuthProvider = ({ children }: IAuthContexProps) => {
       setLoading(false);
     };
     loadUser();
-  }, []);
+  }, [isShowModalPet, isOpenModalSupplies]);
 
   const loginUser = async (data: IUserLogin) => {
     loginUsers(data)
       .then((res) => {
-        const { user: userReponse, accessToken } = res;
+        const { user: userResponse, accessToken } = res;
         api.defaults.headers = {
           Authorization: `bearer ${accessToken}`,
         } as ICommonHeaderProperties;
-        setUser(userReponse);
+
+        setUser(userResponse);
         setIsLogged(true);
-        toast.success("Usuário logado com sucesso!", {
+
+        toast.success("Login realizado com sucesso!", {
           autoClose: 900,
           theme: "dark",
         });
-        navigate("/profile", { replace: true });
-        localStorage.setItem("@AuqMia:token", accessToken);
 
-        localStorage.setItem("@AuqMia:id", `${userReponse.id}`);
+        navigate("/profile", { replace: true });
+
+        localStorage.setItem("@AuqMia:token", accessToken);
+        localStorage.setItem("@AuqMia:id", `${userResponse.id}`);
       })
       .catch((err) =>
-        toast.error("Senha ou email incorreto", {
+        toast.error("Senha ou email incorreto.", {
           autoClose: 900,
           theme: "dark",
         })
       );
   };
+
   const backProfile = () => {
     navigate("/");
     localStorage.removeItem("@AuqMia:token");
@@ -117,6 +147,14 @@ const AuthProvider = ({ children }: IAuthContexProps) => {
       .catch((err) => console.log(err));
   };
 
+  /*   const getSupplies = async () => {
+    await getSuppliesApi()
+      .then((res) => {
+        setListSupplies(res);
+      })
+      .catch((err) => console.log(err));
+  }; */
+
   const deleteAnimal = async (id: string) => {
     const token = localStorage.getItem("@AuqMia:token");
 
@@ -126,21 +164,74 @@ const AuthProvider = ({ children }: IAuthContexProps) => {
     await getAnimals();
   };
 
+  const registerPet = (data: {}) => {
+    const token = localStorage.getItem("@AuqMia:token");
+    const id = localStorage.getItem("@AuqMia:id");
+    const req = { userId: id, ...data };
 
-  const updateUser = async (value: IUpdateUser) => {
-   await upDateUserApi(value)
-   .then((res) => {
-    console.log(res)
-    setModalUpdateUser(false)
-    toast.success("Usuario atualizado com sucesso!")
-   })
-   .catch((err) => {
-    toast.error("Erro ao atualizar!")
-   })
-   
-  }
+    if (token) {
+      api
+        .post("/animals", req, {
+          headers: { Authorization: `Bearer ${token}` },
+        })
+        .then((res) => {
+          toast.success("Animal cadastrado com sucesso!", {
+            autoClose: 900,
+            theme: "dark",
+          });
+          setIsShowModalPet(!isShowModalPet);
+        })
+        .catch((err) => {
+          toast.error("Aconteceu algum erro, verefique os dados!", {
+            autoClose: 900,
+            theme: "dark",
+          });
+        });
+    }
+  };
 
+  const updateUser = async (data: any) => {
+    await upDateUserApi(data)
+      .then((res: any) => {
+        console.log(res);
+        setUser(res);
+        setModalUpdateUser(!modalUpdateUser);
+        toast.success("Usuário atualizado com sucesso!", {
+          autoClose: 900,
+          theme: "dark",
+        });
+      })
+      .catch((err) => {
+        toast.error("Erro ao atualizar!", {
+          autoClose: 900,
+          theme: "dark",
+        });
+      });
+  };
 
+  const registerUser = (data: IUserRegister) => {
+    const { confirm_password, state, district, city, ...restData } = data;
+    const userData = {
+      address: { state: state.toUpperCase(), city, district },
+      ...restData,
+    };
+    api
+      .post("/register", userData)
+      .then((res) => {
+        toast.success("Usuário cadastrado com sucesso!", {
+          autoClose: 900,
+          theme: "dark",
+        });
+        navigate("/login", { replace: true });
+      })
+      .catch((err) => {
+        console.log(err);
+        toast.error("Não foi possível fazer o cadastro!", {
+          autoClose: 900,
+          theme: "dark",
+        });
+      });
+  };
 
   return (
     <AuthContext.Provider
@@ -161,7 +252,15 @@ const AuthProvider = ({ children }: IAuthContexProps) => {
         deleteAnimal,
         modalUpdateUser,
         setModalUpdateUser,
-        updateUser
+        updateUser,
+        registerUser,
+        lisAnimalsUser,
+        registerPet,
+        setIsShowModalPet,
+        isShowModalPet,
+        listSupplies,
+        isOpenModalSupplies,
+        setIsOpenModalSupplis,
       }}
     >
       {children}
